requirements.txt
Django>=4.0
djangorestframework
huey
redis